Fix missing rewrites
Summary:
There's a potential issue where if the only XHP in a file is at the very end it
won't get rewritten.

The issue was that T_RETURN expects a expr_without_variable which is a subset of
expr. Changes to fastpath.re fixes a potential issue where if XHP appears at the
VERY end of a file it won't get picked up.

Reviewed By: dcorson

Test Plan:
The usual inadequate testing of very sensitive code changes. `make test` would be useful.

Revert: OK

DiffCamp Revision: 69833