Fix multiple attribute stealing
Summary:
There's a bug right now where if you steal attributes you have to do them all on the same line or there's a parse error. This isn't a parse error in XHP, but a problem with the generated code. Stripping line number information from the identifier fixes this, which is the desired behavior anyway because __xhpAttributeDeclaration should be one 1 line.

Reviewed By: kdelong

Test Plan:
attributes.phpt

Revert: OK

DiffCamp Revision: 80761