Fix invalid line numbers on syntax error
Summary:
If you encounter a syntax error via a mismatched tag, the line number it gives you may be messed up. This is because we only mark the parser as terminated but let it continue its work. This keeps eating up tokens and advancing the line number. So now we don't let the scanner continue if the parser has been marked as terminated, thus the line number remains where it was when we found the error.

Test Plan:
Unit tests.

Reviewed By: kdelong