Rename :x:base to :xhp
Summary:
We did this internally a while ago, and some tooling assumes
that `:xhp` is the root of the inheritance heirarchy. Rename externally
as well, for maximum compatibility between said tools and external code.
Leave a shim around as `:x:base` as to try to be minimally invasive to
existing code.

Test Plan:
`git grep x:base` shows only the shim remaining. Run this code,
see no errors:

{P7197267}

Reviewers: sparker

Reviewed By: sparker

CC: glevi, ptarjan, joelm

Differential Revision: https://phabricator.fb.com/D1217524

Task ID: 3911343