Add INI Property to Control Global XHP Namespacing
Add a new ini property, “force_global_namespace”, which defaults to
true, to control whether XHP elements should always be instantiated on
the global namespace. Originally this was always done automatically if
your PHP instance was greater than 5.3, but now you can control whether
you want it done or not.