Hackify XHP Core
Hackifies XHP, pure and simple. The big omission is the return type on
getAttribute(). Because it is mixed, if you knew it was a string and
were appending it to another string, XHP could complain that mixed is
not a compatible type with string in strict mode. Just leaving it off
lets the compiler ignore it.