Implement ->: syntax for accessing XHP attributes.
HHVM added syntactic sugar for accessing XHP attributes in facebook/hhvm@38669653bd5449c6d55afc2bed5364631579265b. This syntax essentially maps down to calling getAttribute() on the object. It's a bit strange that it relies on the "userspace" calls, but there's probably no (simple) better way.

Anyways, this adds the same syntax to the Zend PHP extension, plus a test to verify that it works.