RFC: Implement children declarations as a trait
Currently serializes to the same format as HackC, to allow migrating to
the new syntax with no changes to validation.

Some normalization is needed, as HackC emits some overly-complex
structures - e.g. with `$expr = tuple(EXACTLY_ONE, EXPRESSION, $inner_expr)`
can be simplified to `$expr = $inner_expr`.

It is not currently in a form suitable for use - the trait I've added
merely confirms that the two forms of children declaration are
semanticaly equivalent.

I think that's a meaningful first step; if we're happy for this, we can
merge it, start on migration tools, then add an alternative trait to
allow removal of the legacy syntax.

This is very much a draft API; open to a complete rewrite if desired :)

refs #212