Generate all remaining tags without types
I have used a dumb script to generate this.
I feed it the Attributes: block from the spec.
It breaks it into attribute names and filters
out a list of deprecated and already inherited things.
The type is set to `unknown`.
In a later commit, I'll replace the unknowns with actual types.