Clean up dependencies of xhp-lib
This composer.json did not mention that we depended on the hsl.
The hsl came for free with type-assert.
We don't need TypeAssert, since we don't depend on varrayness.
This also improves perf, since we are not spinning in
a loop validating that all elements are mixed, which they always are.